requirements.txt
python-telegram-bot
httpx
APScheduler
pydantic
python-dotenv
playwright
